refactor ($step3): 피드백 반영
- LottoWinning -> Rank로 파일명 변경
