+++ src/main/java/step3/domain/LottoStatistic.java
@@ -9,7 +9,8 @@ import java.util.Map;
 public class LottoStatistic {
 	private final Lotto winningLotto;
 	private final int winningBonus;
-	private static final Map<LottoWinning, Integer> lottoStat = new HashMap<>();
+
+	private static final Map<Rank, Integer> lottoStat = new HashMap<>();
 	private static final String INIT_SUM = "0";
 	private static final int SCALE = 2;
 
@@ -17,8 +18,7 @@ public class LottoStatistic {
 		this.winningLotto = winningLotto;
 		this.winningBonus = winningBonus;
 	}
-
-	public Map<LottoWinning, Integer> calcLottoResult(List<Lotto> lottos) {
+	public Map<Rank, Integer> calcLottoResult(List<Lotto> lottos) {
 		for (Lotto lotto : lottos) {
 			calcWinningNumberCount(lotto);
 		}
@@ -27,18 +27,18 @@ public class LottoStatistic {
 
 	private void calcWinningNumberCount(Lotto lotto) {
 		int count = lotto.getMatchCount(winningLotto);
-		LottoWinning winning = LottoWinning.valueOf(count, checkBonus(lotto));
+		Rank winning = Rank.valueOf(count, checkBonus(lotto));
 		lottoStat.put(winning, lottoStat.getOrDefault(winning, 0)+1);
 	}
 
 	private boolean checkBonus(Lotto lotto) {
 		int count = lotto.getMatchCount(winningLotto);
-		return count == LottoWinning.FIVE.getMatchingCount() && lotto.containNumber(new LottoNumber(winningBonus));
+		return count == Rank.FIVE.getMatchingCount() && lotto.containNumber(new LottoNumber(winningBonus));
 	}
 
 	public BigDecimal calcYield(int price) {
 		BigDecimal sum = new BigDecimal(INIT_SUM);
-		for (LottoWinning winning : lottoStat.keySet()) {
+		for (Rank winning : lottoStat.keySet()) {
 			int s = lottoStat.get(winning) * winning.getReward();
 			sum = sum.add(BigDecimal.valueOf(s));
 		}
